Collapse AbstractBoard into ChessBoard default methods
Every method on AbstractBoard was a stateless derivation expressible
in terms of the ChessBoard interface — the textbook case for default
methods. Java 9+ also lets us keep the private isGameDraw() helper
inside the interface as a private method.

- ChessBoard extends EnumConstants (preserves unqualified enum
  constants for implementations that rely on them).
- 12 default methods added: getCastlingRight(Side),
  canClaimFiftyMoveRule, canClaimThreefoldRepetitionRule,
  isDeadPositionQuick/Full, isUnwinnableQuick/Full,
  calculateInsufficientMaterial, isGameEnd,
  getFullMoveNumberForNextHalfMove, getLegalMovesSan, getLegalMovesUci.
- Board and LibraryCarlosBoard now implements ChessBoard directly.
- AbstractBoard.java deleted.

Side benefit: CHA methods (isUnwinnableQuick/Full, isDeadPositionQuick/Full)
now appear directly on Board / ChessBoard IDE completion. This
supersedes the matching sub-item under Task 5.

950/950 tests still green.
