--- src/main/java/com/dlb/chess/common/AbstractBoard.java
@@ -1,158 +1,0 @@
-package com.dlb.chess.common;
-
-import java.util.Set;
-import java.util.TreeSet;
-
-import com.dlb.chess.board.enums.CastlingRight;
-import com.dlb.chess.board.enums.Side;
-import com.dlb.chess.common.constants.EnumConstants;
-import com.dlb.chess.common.enums.InsufficientMaterial;
-import com.dlb.chess.common.interfaces.ChessBoard;
-import com.dlb.chess.common.model.MoveSpecification;
-import com.dlb.chess.common.ucimove.utility.UciMoveUtility;
-import com.dlb.chess.unwinnability.full.UnwinnableFullAnalyzer;
-import com.dlb.chess.unwinnability.full.enums.DeadPositionFull;
-import com.dlb.chess.unwinnability.full.enums.UnwinnableFull;
-import com.dlb.chess.unwinnability.quick.UnwinnableQuickAnalyzer;
-import com.dlb.chess.unwinnability.quick.enums.DeadPositionQuick;
-import com.dlb.chess.unwinnability.quick.enums.UnwinnableQuick;
-import com.google.common.collect.ImmutableSet;
-
-public abstract class AbstractBoard implements ChessBoard, EnumConstants {
-
-  @Override
-  public ImmutableSet<String> getLegalMovesSan() {
-    final Set<String> result = new TreeSet<>();
-
-    for (final MoveSpecification moveSpecification : getPossibleMoveSpecificationSet()) {
-      this.performMove(moveSpecification);
-      result.add(getSan());
-      this.unperformMove();
-    }
-    return NonNullWrapperCommon.copyOfSet(result);
-  }
-
-  @Override
-  public ImmutableSet<String> getLegalMovesUci() {
-    final Set<String> result = new TreeSet<>();
-    final Side havingMove = getHavingMove();
-    for (final MoveSpecification moveSpecification : getPossibleMoveSpecificationSet()) {
-      final String uci = UciMoveUtility.convertMoveSpecificationToUci(havingMove, moveSpecification).text();
-      result.add(uci);
-    }
-    return NonNullWrapperCommon.copyOfSet(result);
-  }
-
-  @Override
-  public CastlingRight getCastlingRight(Side havingMove) {
-
-    return switch (havingMove) {
-      case WHITE -> getDynamicPosition().castlingRightWhite();
-      case BLACK -> getDynamicPosition().castlingRightBlack();
-      case NONE -> throw new IllegalArgumentException();
-      default -> throw new IllegalArgumentException();
-    };
-  }
-
-  @Override
-  public DeadPositionQuick isDeadPositionQuick() {
-    final UnwinnableQuick unwinnableWhite = UnwinnableQuickAnalyzer.unwinnableQuick(this, Side.WHITE);
-    final UnwinnableQuick unwinnableBlack = UnwinnableQuickAnalyzer.unwinnableQuick(this, Side.BLACK);
-
-    if (unwinnableWhite == UnwinnableQuick.UNWINNABLE && unwinnableBlack == UnwinnableQuick.UNWINNABLE) {
-      return DeadPositionQuick.DEAD_POSITION;
-    }
-
-    if (unwinnableWhite == UnwinnableQuick.WINNABLE && unwinnableBlack == UnwinnableQuick.WINNABLE) {
-      return DeadPositionQuick.NON_DEAD_POSITION;
-    }
-
-    return DeadPositionQuick.POSSIBLY_NON_DEAD_POSITION;
-  }
-
-  @Override
-  public DeadPositionFull isDeadPositionFull() {
-    final UnwinnableFull unwinnableWhite = UnwinnableFullAnalyzer.unwinnableFull(this, Side.WHITE).unwinnableFull();
-    if (unwinnableWhite == UnwinnableFull.WINNABLE) {
-      return DeadPositionFull.NON_DEAD_POSITION;
-    }
-
-    final UnwinnableFull unwinnableBlack = UnwinnableFullAnalyzer.unwinnableFull(this, Side.BLACK).unwinnableFull();
-    if (unwinnableBlack == UnwinnableFull.WINNABLE) {
-      return DeadPositionFull.NON_DEAD_POSITION;
-    }
-
-    if (unwinnableWhite == UnwinnableFull.UNWINNABLE && unwinnableBlack == UnwinnableFull.UNWINNABLE) {
-      return DeadPositionFull.DEAD_POSITION;
-    }
-
-    return DeadPositionFull.UNDETERMINED;
-  }
-
-  @Override
-  public UnwinnableQuick isUnwinnableQuick(Side side) {
-    return UnwinnableQuickAnalyzer.unwinnableQuick(this, side);
-  }
-
-  @Override
-  public UnwinnableFull isUnwinnableFull(Side side) {
-    return UnwinnableFullAnalyzer.unwinnableFull(this, side).unwinnableFull();
-  }
-
-  @Override
-  public InsufficientMaterial calculateInsufficientMaterial() {
-    if (isInsufficientMaterial()) {
-      return InsufficientMaterial.BOTH;
-    }
-    if (isInsufficientMaterial(WHITE)) {
-      return InsufficientMaterial.WHITE_ONLY;
-    }
-    if (isInsufficientMaterial(BLACK)) {
-      return InsufficientMaterial.BLACK_ONLY;
-    }
-    return InsufficientMaterial.NONE;
-  }
-
-  @Override
-  public boolean isGameEnd() {
-    if (isCheckmate()) {
-      return true;
-    }
-    return isGameDraw();
-  }
-
-  private boolean isGameDraw() {
-    return isStalemate() || isDeadPositionQuick() == DeadPositionQuick.DEAD_POSITION || isFivefoldRepetition()
-        || isSeventyFiveMove();
-  }
-
-  @Override
-  public boolean canClaimFiftyMoveRule() {
-    if (isFiftyMove()) {
-      return true;
-    }
-    return canClaimFiftyMoveRuleWithOwnMove();
-  }
-
-  @Override
-  public boolean canClaimThreefoldRepetitionRule() {
-    if (isThreefoldRepetition()) {
-      return true;
-    }
-    return canClaimThreefoldRepetitionRuleWithOwnMove();
-  }
-
-  @Override
-  public int getFullMoveNumberForNextHalfMove() {
-    if (isFirstMove()) {
-      return getInitialFenFullMoveNumber();
-    }
-    return switch (getHavingMove()) {
-      case BLACK -> getFullMoveNumber();
-      case WHITE -> getFullMoveNumber() + 1;
-      case NONE -> throw new IllegalArgumentException();
-      default -> throw new IllegalArgumentException();
-    };
-  }
-
-}
